Components: Cleaning up streamedtables, coupling normaltable
extracted `TableBody` and `TableHeader`, I am not turning them into
component stories because their effects are obvious when you look at the
streamed or normal tables and it is difficult to insert a tanstack table
there.

Removed streamedtables.d.ts
